removed orginal level structure
