--- src/level/blocks/mod.rs
@@ -1,4 +1,0 @@
-pub mod air;
-pub mod blocks;
-pub mod full_block;
-pub mod meshable;
