--- src/level/prefabs/mod.rs
@@ -1,1 +1,0 @@
-pub mod prefabs;
