--- src/level/prefabs/prefabs.rs
@@ -1,2 +1,0 @@
-pub enum Prefab{
-}
